Ruleset::loadFile() - wb's fix for double terror
change a bunch of ints to size_ts, etc.

+++ src/Savegame/AlienStrategy.h
@@ -42,7 +42,7 @@ private:
 	std::map<std::string, WeightedOptions*> _regionMissions; // The chances of each mission type for each region.
 
 	std::map<std::string, int> _missionRuns;
-	std::map<std::string, std::vector<std::pair<std::string, int> > > _missionLocations;
+	std::map<std::string, std::vector<std::pair<std::string, size_t> > > _missionLocations;
 
 	/// Disable copy and assignments.
 	AlienStrategy(const AlienStrategy&);
@@ -82,13 +82,13 @@ private:
 		void addMissionLocation(
 				const std::string& id,
 				const std::string& region,
-				int zone,
+				size_t zone,
 				size_t track);
 		/// Checks if a given mission location has been attacked already.
 		bool validMissionLocation(
 				const std::string& id,
 				const std::string& region,
-				int zone);
+				size_t zone);
 		/// Checks that a given region appears in the strategy table.
 		bool validMissionRegion(const std::string& region) const;
 };
